requirements.txt
Cython>=0.29.21
numpy>=1.18.1
pandas>=1.4.0
pyjnius>=1.4.0
scikit-learn>=0.22.1
scipy>=1.4.1
tqdm
transformers>=4.6.0
sentencepiece>=0.1.95
nmslib>=2.1.1
onnxruntime>=1.8.1
lightgbm>=3.3.2
spacy>=3.2.1
pyyaml
openai>=1.0.0
tiktoken>=0.4.0
pyarrow>=15.0.0
pillow>=10.2.0
open-clip-torch>=2.24.0